Tests: Attack: Set up combat tests of the AttackLegal class.

+++ deploy/tests/TestAccountCreateAndDestroy.php
@@ -4,24 +4,27 @@ require_once(LIB_ROOT.'base.inc.php');
 
 class TestAccountCreateAndDestroy{
 	public static $test_email = 'testphpunit@example.com';
+	public static $alt_test_email = 'testphpunit2@example.com';
 	public static $test_password = 'password';
 	public static $test_ninja_name = 'phpunit_ninja_name';
+	public static $alt_test_ninja_name = 'phpunit_alt_ninja';
 
 	// Library for creating and destroying test-only accounts, for use in their various ways in testing.
 	public static function purge_test_accounts($test=null){
 	    $test_ninja_name = $test? $test : TestAccountCreateAndDestroy::$test_ninja_name;
-	    $active_email = 'testphpunit@example.com';
+	    $active_email = TestAccountCreateAndDestroy::$test_email;
+	    $alt_active_email = TestAccountCreateAndDestroy::$alt_test_email;
 	    $aid = get_char_id(TestAccountCreateAndDestroy::$test_ninja_name);
 	    query('delete from players where player_id in 
 	        (select player_id from players left join account_players on _player_id = player_id 
 	        	left join accounts on _account_id = account_id 
-	            where active_email = :active_email or account_identity= :ae2 or players.uname = :uname)', 
-	        array(':active_email'=>$active_email, ':ae2'=>$active_email, ':uname'=>$test_ninja_name)); // Delete the players
+	            where active_email = :active_email or account_identity= :ae2 or players.uname = :uname or active_email = :alt_active_email or account_identity = :alt_active_email2)', 
+	        array(':active_email'=>$active_email, ':ae2'=>$active_email, ':uname'=>$test_ninja_name, ':alt_active_email'=>$alt_active_email, ':alt_active_email2'=>$alt_active_email)); // Delete the players
 	    query('delete from account_players where _account_id in (select account_id from accounts 
-	            where active_email = :active_email or account_identity= :ae2)', // Delete the account_players linkage.
-	        array(':active_email'=>$active_email, ':ae2'=>$active_email));
-	    $query = query('delete from accounts where active_email = :active_email or account_identity= :ae2', 
-	    	array(':active_email'=>$active_email, ':ae2'=>$active_email)); // Finally, delete the test account.
+	            where active_email = :active_email or account_identity= :ae2 or active_email = :alt_active_email or account_identity = :alt_active_email2)', // Delete the account_players linkage.
+	        array(':active_email'=>$active_email, ':ae2'=>$active_email,  ':alt_active_email'=>$alt_active_email, ':alt_active_email2'=>$alt_active_email));
+	    $query = query('delete from accounts where active_email = :active_email or account_identity= :ae2 or active_email = :alt_active_email or account_identity = :alt_active_email2', 
+	    	array(':active_email'=>$active_email, ':ae2'=>$active_email,  ':alt_active_email'=>$alt_active_email, ':alt_active_email2'=>$alt_active_email)); // Finally, delete the test account.
 	    return ($query->rowCount() > 0);
 	    
 	    /*
@@ -63,6 +66,34 @@ class TestAccountCreateAndDestroy{
 		}
 		$_SERVER['REMOTE_ADDR']=$previous_server; // Reset remote addr to whatever it was before.
 		$char_id = get_char_id(TestAccountCreateAndDestroy::$test_ninja_name);
+		return $char_id;
+	}
+
+	// Create a separate, second testing account
+	public static function create_alternate_testing_account($confirm=false){
+		if((bool)get_char_id(TestAccountCreateAndDestroy::$alt_test_ninja_name)){
+			throw new Exception('Test user found ['.$found.'] with name ['.TestAccountCreateAndDestroy::$alt_test_ninja_name.'] already exists');
+		}
+		// Create test user, unconfirmed, whatever the default is for activity.
+		$preconfirm = true;
+		$confirm = rand(1000,9999); //generate confirmation code
+
+		// Use the function from lib_player
+		$player_params = array(
+			'send_email'    => TestAccountCreateAndDestroy::$alt_test_email
+			, 'send_pass'   => TestAccountCreateAndDestroy::$test_password
+			, 'send_class'  => 'dragon'
+			, 'preconfirm'  => true
+			, 'confirm'     => $confirm
+			, 'referred_by' => 'ninjawars.net'
+		);
+		ob_start(); // Skip extra output
+		$error = create_account_and_ninja(TestAccountCreateAndDestroy::$alt_test_ninja_name, $player_params);
+		ob_end_clean();
+		if($confirm){
+			$confirmed = confirm_player(TestAccountCreateAndDestroy::$alt_test_ninja_name, false, true); // name, no confirm #, just autoconfirm.
+		}
+		$char_id = get_char_id(TestAccountCreateAndDestroy::$alt_test_ninja_name);
 		return $char_id;
 	}
